Add publishing indexes to build-docs workflow
The workflow in the GitHub actions "build-docs.yml" file has been updated to include a new job that publishes the indexes. This job starts after the build, test, and deploy and will unzip the doc artifacts, use the JetBrains publisher to update the Algolia index, and log the output. The changes aim to automate index publishing, thereby simplifying the documentation update process.